Backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
pillow
pypdf2
sqlalchemy
jinja2
aiofiles
pdf2image      
